fix(tests): major upgrade to Nest.js v11 upgrade
Used ncu to bump all dependencies to latest versions. The Nest.js upgrade
from v10 to v11 introduced breaking changes that necessitated some code changes:

- There was a circular dependency between the EnvelopeService and PaginationService
as a consequence of importing the EnvelopeDto via barrel import. Changed
pagination.service.ts to import EnvelopeDto directly from ../envelope/dto
instead of ../envelope.

- Changed all Nest TestingModule instances in spec.ts files to async retrieve
services using module.resolve instead of module.get. This is a consequence of
the TaxiiLoggerService being request-scoped and it being a dependency
throughout the majority (if not all) of the modules. There was also a
missing import (TaxiiConfigModule) in pagination.service.spec.ts.

- Removed the uuid and @types/uuid packages, opting instead to rely on the
standard library for uuid generation. The upgrade of uuid from v11 to v13
broke compatibility with our app due to a transpilation issue (I think related
to CJS vs ESM but unclear). For our uses, the built-in crypto.randomUUID
function works just as well.