fix(hydrate): enable pulling multiple versions of a STIX collection
There was a bug where, if the Workbench sidecar contained multiple versions
of the same collection (e.g., ATT&CK v17.1 and v18.1), then the TAXII server
would only sync whichever version of the collection that was observed first.

This change modifies the WorkbenchRepository to resolve this issue. The
?versions=all query paramater is now used to retrieve all available
collections. In addition, when retrieving individual objects via
ObjectRepository.findOne, we were only retrieving the latest copy. The
active filter has been removed, since filtering is handled by the filter
service.